#16 - moved controller to new package
