+++ src/main/java/de/bayern/gdi/gui/controller/ButtonBarController.java
@@ -1,4 +1,4 @@
-package de.bayern.gdi.gui;
+package de.bayern.gdi.gui.controller;
 
 import de.bayern.gdi.model.DownloadStep;
 import de.bayern.gdi.processor.ConverterException;
@@ -21,7 +21,7 @@ import javax.inject.Inject;
 import java.io.File;
 import java.io.IOException;
 
-import static de.bayern.gdi.gui.Controller.USER_DIR;
+import static de.bayern.gdi.gui.controller.Controller.USER_DIR;
 
 /**
  * @author <a href="mailto:goltz@lat-lon.de">Lyn Goltz </a>
