+++ src/main/java/de/bayern/gdi/gui/controller/ServiceSelectionController.java
@@ -1,7 +1,9 @@
-package de.bayern.gdi.gui;
+package de.bayern.gdi.gui.controller;
 
+import de.bayern.gdi.gui.ServiceModel;
 import de.bayern.gdi.services.Atom;
 import de.bayern.gdi.services.Service;
+import de.bayern.gdi.services.ServiceType;
 import de.bayern.gdi.services.WFSMeta;
 import de.bayern.gdi.services.WFSMetaExtractor;
 import de.bayern.gdi.utils.DownloadConfig;
@@ -34,6 +36,10 @@ import java.net.MalformedURLException;
 import java.net.URISyntaxException;
 import java.net.URL;
 import java.util.List;
+
+import static de.bayern.gdi.services.ServiceType.ATOM;
+import static de.bayern.gdi.services.ServiceType.WFS_ONE;
+import static de.bayern.gdi.services.ServiceType.WFS_TWO;
 
 /**
  * @author <a href="mailto:goltz@lat-lon.de">Lyn Goltz </a>
