+++ src/resources/button-bar.fxml
@@ -10,7 +10,7 @@
       minHeight="-Infinity" prefHeight="54.0"
       prefWidth="590.0" VBox.vgrow="NEVER" xmlns="http://javafx.com/javafx/8"
       xmlns:fx="http://javafx.com/fxml/1"
-      fx:controller="de.bayern.gdi.gui.ButtonBarController">
+      fx:controller="de.bayern.gdi.gui.controller.ButtonBarController">
     <children>
         <Button fx:id="buttonDownload"
                 mnemonicParsing="false"
