+++ src/resources/download-client.fxml
@@ -14,7 +14,7 @@
 
 <VBox prefHeight="622.0" prefWidth="890.0" xmlns="http://javafx.com/javafx/8"
       xmlns:fx="http://javafx.com/fxml/1"
-      fx:controller="de.bayern.gdi.gui.Controller">
+      fx:controller="de.bayern.gdi.gui.controller.Controller">
     <children>
         <fx:include fx:id="menuBar" source="menu-bar.fxml" />
         <SplitPane dividerPositions="0.3311036789297659" VBox.vgrow="ALWAYS">
