+++ src/resources/menu-bar.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.control.MenuItem?>
 <MenuBar xmlns="http://javafx.com/javafx/8"
          xmlns:fx="http://javafx.com/fxml/1"
-         fx:id="menuBar" fx:controller="de.bayern.gdi.gui.MenuBarController">
+         fx:id="menuBar" fx:controller="de.bayern.gdi.gui.controller.MenuBarController">
     <menus>
         <Menu text="%menu.file" fx:id="menuFile">
             <MenuItem text="%menu.load_config" onAction="#handleLoadConfig"
