+++ src/resources/processing-chain.fxml
@@ -11,7 +11,7 @@
 <VBox fx:id="chainContainer"
       xmlns="http://javafx.com/javafx/8"
       xmlns:fx="http://javafx.com/fxml/1"
-      fx:controller="de.bayern.gdi.gui.ProcessingChainController">
+      fx:controller="de.bayern.gdi.gui.controller.ProcessingChainController">
     <children>
         <CheckBox
             fx:id="chkChain"
